refactor(settings-hub): compose via index.ts
Create lib/codex-manager/settings-hub/index.ts as the composition
root for the split settings hub. It owns:

- configureUnifiedSettings (public entry)
- promptSettingsHub (internal hub prompt)
- __testOnly barrel re-exporting 16 internals sourced from
  backend.ts, dashboard.ts, experimental.ts, and shared.ts

lib/codex-manager/settings-hub.ts is now a 9-line back-compat
re-export stub pointing at ./settings-hub/index.js. Keeping the stub
avoids rewriting two runtime consumers (lib/codex-manager.ts and
lib/codex-manager/auth-commands.ts) and the existing vi.mock in
test/codex-manager-auth-commands.test.ts, and preserves the import
path used by test/settings-hub-utils.test.ts.

No behavior change: all 302 targeted tests (settings-hub +
codex-manager) remain green. Full suite verification follows in the
final commit.

Refs: AUDIT-M24, G-01, JN-03
